high-level-api: when no device selected switch to u2f, temporarily
For WebUSB fallback to u2f if the user closes the device selector without selecting
a device as the user might not have been able to select his device due to the Nano X
currently not being discoverable via WebUSB in Windows.
This fallback also temporarily servers Linux users which have not updated their udev
rules yet.

The fallback is just temporary and to be removed once WebUSB with Nano X works on
Windows or WebHID is more broadly available.