rollup: switch to @rbnlffl/rollup-plugin-eslint
Previously used rollup-plugin-eslint is not maintained anymore and is stuck
at eslint 6. It also imported eslint configs of dependencies, even though
node_modules was ignored. This leads to problems for some dependencies
required for bitcoinjs-lib which have eslint files with eslint 7 features.

The new plugin is also able to handle typescript files out of the box without
the need for our previous fixedEslint fix.