RequestBitcoin: disallow using testnet and mainnet apps interchangeably
Operating on testnet paths is generally allowed also for the Bitcoin mainnet app
and retrieved public keys (and thus also computed extended keys) and signed
transactions are identical to the testnet app, however addresses displayed or
generated by the Ledger are in mainnet format, regardless of whether a testnet
path is specified. The testnet app since version 1.4.6 shows a warning when
accessing mainnet paths. For these reason we generally block using the Bitcoin
mainnet and testnet apps interchangeably.