high-level-api: simplify getWalletId logic
As the connection for transport types other than U2F and WebAuthn is
maintained and not reset after each request, there is no reduced cost
in calling getWalletId in parallel to another call, as the connection
info can be re-used by another request when calling getWalletId before
or getWalletId can use the connection info of a previous request.

For U2F and WebAuthn the old implementation would still save one
checkCoinAppConnection call when calling getWalletId in parallel to
another request. However, the saving is not that relevant as getWalletId
is typically only used once at wallet import and due to the deprecation
of these transport types, it's not worth the additional complexity
anymore.

Now getWalletId can not be called in parallel to other requests
anymore, which is more consistent behavior.

This code simplification is also in preparation in making the wallet id
computation optional.