Request: make a cached _coinAppConnection available to all requests
Previously we cached it only within RequestGetWalletIdNimiq and RequestGetWalletIdBitcoin, but after
updating the ledger libraries, we'll also want it for all Bitcoin requests for checking the version of the
connected app for deciding which api version to use.

Additionally, now make sure to update the current _coinAppConnection and _coinAppConnection.walletId before
potentially throwing, if it's not the expected one.