Update ledger packages which now provide support for Bitcoin apps before and after 2.0
This commit does however not make the appropriate changes yet to make our API compatible
with the new Bitcoin app's api included in Bitcoin apps starting with 2.0 and mandatory
with 2.1. Instead, here we still enforce usage of the old api and only make the changes
to make the code compile and still work with the old api.