rollup: replace rollup-plugin-polyfill-node by individual polyfills
Polyfill node builtins manually via individual npm packages and manual shims, to be able to
update them manually to maintained versions. Notably, the buffer polyfill has been updated
this way, for compatibility with the newest @ledgerhq/hw-app-btc, which uses readUint8() in
BtcNew.signMessage without specifying an offset, which is not supported yet by the polyfill
bundlded with rollup-plugin-polyfill-node.

In order to properly handle the circular dependencies of the now updated readable-stream
package, @rollup/plugin-commonjs had to be updated to at least version 22.0.0 to include
https://github.com/rollup/plugins/pull/1038 to fix https://github.com/rollup/rollup/issues/1507.

Additionally, rollup and @rollup/plugin-node-resolve had been updated to the minimum version
compatible with the updated @rollup/plugin-commonjs.