Update rollup and rollup plugins
Also set the package type to module to signal to rollup that rollup.config.js is an es module.
At the same time rename .eslintrc.js to .eslintrc.cjs as it's not an es module, but a commonjs module.

As unfortunately rollup-plugin-sourcemaps has not been updated in a while and uses an old version of
@rollup/pluginutils which is incompatible with newer rollup versions, its dependency has been manually
overwritten in package.json via npm's "overrides" feature and yarn's "resolutions" feature.