demo: add buffer utils
Replace usage of BufferUtils and SerialBuffer from Nimiq core with our own buffer utils.
This removes the need to load the Nimiq library in some places in the demo page only for
its BufferUtils and SerialBuffer, which in turn allows for better testing of the loading
behavior of the api itself, as opposed to the demo page.
This is also in preparation for supporting the Albatros library, which does not provide
the BufferUtils and SerialBuffer classes anymore.