Support Nimiq Albatross core-rs library
Makes the API compatible with the legacy Nimiq core and the Albatross core-rs library.
Notably, this change makes the api able to use the appropriate library for internal calculations
and accept input and return output objects of the appropriate library. However, this commit does
not contain required updates yet for communicating with the new Albatross compatible version of
the Ledger app, or changes to contract data on the demo page, and does not provide support for
newly introduced sender data and new transaction types on the demo page yet.

The API now supports both library versions with the idea of being compatible with Nimiq PoW and
Nimiq Albatross, and the associated versions of the Ledger app, to maintain a common code base.

Notably, this change is completely free of breaking changes until the mainnet fully migrates to
Albatross, at which point the idea is to remove support for the legacy library.