demo: on demo page don't preload Nimiq wasm
If Nimiq needs to be loaded separately on the demo page, don't preload its wasm.
As the APIs are used from separate pre-built bundles, the cached load promises
between the API bundles and the demo page bundle are different, and thus the
wasm would be preloaded again, which is then not actually used anymore by the
browser if it already loaded the wasm, resulting in a warning in the console.
Additionally, without the preload, the browser apparently doesn't even fetch the
wasm again: there's no second entry in the network tab of the dev tools.