high-level-api,low-level-api: make Albatross the default Nimiq version
Previously, it was NimiqVersion.LEGACY to avoid introducing a breaking change in the API
methods. However, releasing a new major version nonetheless is probably nicer, given the
vast amount of changes (even if defaults were previously backwards compatible), and to
make users of Nimiq Legacy aware of upcoming required changes for Nimiq Albatross.